Add Display monthly sales report feature
If the start month and end month are the same, it will display the detail sales report of that month, while if they are different, it will display all the months in the range with year and the total sales of that month. User can double-click the month to view the details.