Update frmSalesReportM, frmVisualize, frmVRevenue and frmVTop
Set condition for the Visualize and Export buttons to display in frmSalesReportM. Adjusted the arrangement of elements on the frmVisualize and added FormClosing event to it. Changed the line color in the graph in frmVRevenue and added hover tooltip to the graph. Updated frmVTop so that it can display all subgroup for different month and added hover tooltip to the barchart. FormClosing event also added to frmVTop.