Fix bugs
In frmSalesManagement, the "Back" button only show up when the user click the "Edit" button. For frmSalesReportD and frmSalesReportW, the "Visualize" button is removed and the "Export" button only show up when there is record in the datagrid. For frmVRevenue, the majorGrid for x-axis and y-axis is removed so that it is easier to see.